remove other references
