+++ internal/api-key/command_store.go
@@ -7,9 +7,9 @@ import (
 
 	pcmd "github.com/confluentinc/cli/v3/pkg/cmd"
 	"github.com/confluentinc/cli/v3/pkg/config"
-	dynamicconfig "github.com/confluentinc/cli/v3/pkg/dynamic-config"
 	"github.com/confluentinc/cli/v3/pkg/errors"
 	"github.com/confluentinc/cli/v3/pkg/examples"
+	"github.com/confluentinc/cli/v3/pkg/kafka"
 	"github.com/confluentinc/cli/v3/pkg/output"
 	"github.com/confluentinc/cli/v3/pkg/resource"
 )
@@ -67,12 +67,12 @@ func (c *command) store(cmd *cobra.Command, args []string) error {
 		if resourceType != resource.KafkaCluster {
 			return fmt.Errorf(nonKafkaNotImplementedErrorMsg)
 		}
-		cluster, err = dynamicconfig.FindKafkaCluster(c.V2Client, c.Context, clusterId)
+		cluster, err = kafka.FindCluster(c.V2Client, c.Context, clusterId)
 		if err != nil {
 			return err
 		}
 	} else {
-		cluster, err = dynamicconfig.GetKafkaClusterForCommand(c.V2Client, c.Context)
+		cluster, err = kafka.GetClusterForCommand(c.V2Client, c.Context)
 		if err != nil {
 			// Replace the error msg since it suggests flags which are unavailable with this command
 			return errors.NewErrorWithSuggestions(
