+++ internal/kafka/command_cluster_use.go
@@ -6,8 +6,8 @@ import (
 	"github.com/spf13/cobra"
 
 	pcmd "github.com/confluentinc/cli/v3/pkg/cmd"
-	dynamicconfig "github.com/confluentinc/cli/v3/pkg/dynamic-config"
 	"github.com/confluentinc/cli/v3/pkg/errors"
+	"github.com/confluentinc/cli/v3/pkg/kafka"
 	"github.com/confluentinc/cli/v3/pkg/output"
 )
 
@@ -31,7 +31,7 @@ func (c *clusterCommand) newUseCommand() *cobra.Command {
 func (c *clusterCommand) use(_ *cobra.Command, args []string) error {
 	id := args[0]
 
-	if _, err := dynamicconfig.FindKafkaCluster(c.V2Client, c.Context, id); err != nil {
+	if _, err := kafka.FindCluster(c.V2Client, c.Context, id); err != nil {
 		return errors.NewErrorWithSuggestions(
 			fmt.Sprintf(errors.KafkaClusterNotFoundErrorMsg, id),
 			errors.ChooseRightEnvironmentSuggestions,
