+++ internal/kafka/command_topic_produce.go
@@ -16,9 +16,9 @@ import (
 
 	sr "github.com/confluentinc/cli/v3/internal/schema-registry"
 	pcmd "github.com/confluentinc/cli/v3/pkg/cmd"
-	dynamicconfig "github.com/confluentinc/cli/v3/pkg/dynamic-config"
 	"github.com/confluentinc/cli/v3/pkg/errors"
 	"github.com/confluentinc/cli/v3/pkg/examples"
+	"github.com/confluentinc/cli/v3/pkg/kafka"
 	"github.com/confluentinc/cli/v3/pkg/log"
 	"github.com/confluentinc/cli/v3/pkg/serdes"
 )
@@ -115,7 +115,7 @@ func (c *command) produce(cmd *cobra.Command, args []string) error {
 func (c *command) produceCloud(cmd *cobra.Command, args []string) error {
 	topic := args[0]
 
-	cluster, err := dynamicconfig.GetKafkaClusterForCommand(c.V2Client, c.Context)
+	cluster, err := kafka.GetClusterForCommand(c.V2Client, c.Context)
 	if err != nil {
 		return err
 	}
