+++ pkg/keystore/keystore.go
@@ -3,7 +3,7 @@ package keystore
 import (
 	"github.com/confluentinc/cli/v3/pkg/ccloudv2"
 	"github.com/confluentinc/cli/v3/pkg/config"
-	dynamicconfig "github.com/confluentinc/cli/v3/pkg/dynamic-config"
+	"github.com/confluentinc/cli/v3/pkg/kafka"
 )
 
 type ConfigKeyStore struct {
@@ -11,7 +11,7 @@ type ConfigKeyStore struct {
 }
 
 func (c *ConfigKeyStore) HasAPIKey(client *ccloudv2.Client, key, clusterId string) (bool, error) {
-	kcc, err := dynamicconfig.FindKafkaCluster(client, c.Config.Context(), clusterId)
+	kcc, err := kafka.FindCluster(client, c.Config.Context(), clusterId)
 	if err != nil {
 		return false, err
 	}
@@ -22,7 +22,7 @@ func (c *ConfigKeyStore) HasAPIKey(client *ccloudv2.Client, key, clusterId strin
 
 // StoreAPIKey creates a new API key pair in the local key store for later usage
 func (c *ConfigKeyStore) StoreAPIKey(client *ccloudv2.Client, key *config.APIKeyPair, clusterId string) error {
-	kcc, err := dynamicconfig.FindKafkaCluster(client, c.Config.Context(), clusterId)
+	kcc, err := kafka.FindCluster(client, c.Config.Context(), clusterId)
 	if err != nil {
 		return err
 	}
